added graphql files in their own folder
