+++ src/components/Dashboard/components/WorkoutsCard/graphql/CreateWorkoutPlan.ts
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+
+export const CREATE_WORKOUT_PLAN = gql`
+  mutation workoutPlanCreate($name: String!) {
+    createWorkoutPlan(name: $name) {
+      result {
+        id
+        name
+      }
+      error {
+        message
+        name
+      }
+    }
+  }
+`;
