+++ src/components/Dashboard/components/WorkoutsCard/graphql/GetWorkoutPlan.ts
@@ -1,0 +1,10 @@
+import gql from 'graphql-tag';
+
+export const GET_WORKOUT_PLANS = gql`
+  query getWorkoutPlans {
+    getWorkoutPlans {
+      id
+      name
+    }
+  }
+`;
