+++ src/components/Dashboard/components/WorkoutsCard/graphql/index.ts
@@ -1,0 +1,3 @@
+export { GET_WORKOUT_PLANS } from "./GetWorkoutPlan";
+export { CREATE_WORKOUT_PLAN } from "./CreateWorkoutPlan";
+export { UPDATE_ACTIVE_WORKOUT } from "./UpdateActiveWorkout";
